プルリクエストの一覧へのリンクをナビゲーションバーに追加
