+++ renderer/src/containers/PullRequestListContainer/PullRequestListContainer.tsx
@@ -3,10 +3,6 @@ import { usePullRequests } from '../../hooks/usePullRequests';
 import { RepositorySection } from '../../components/RepositorySection';
 import { PullRequest, Repositories } from '../../models';
 
-// 頻度が多いと Github GraphQL API のレート制限に影響するので、投げるクエリのスコアを計算して調整してください。
-// @see: https://docs.github.com/ja/graphql/overview/resource-limitations
-const FETCH_PULL_REQUESTS_INTERVAL = 60 * 1000; // ms
-
 /**
  * プルリクエストをリポジトリ毎にまとめる
  */
@@ -41,12 +37,10 @@ const groupByRepository = (pullRequests: Array<PullRequest>): Repositories => {
 };
 
 export const PullRequestListContainer: FC = () => {
-  const { pullRequests, loading } = usePullRequests({
-    fetchInterval: FETCH_PULL_REQUESTS_INTERVAL,
-  });
+  const { pullRequests, firstLoading } = usePullRequests();
   const repositories = groupByRepository(pullRequests);
 
-  if (loading) {
+  if (firstLoading) {
     return <h1>Loading...</h1>;
   }
