+++ renderer/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './usePullRequests';
 export * from './useSettings';
+export * from './useWatchPullRequests';
